+++ Messages/MessageTypes/ClientDataMessage.cpp
@@ -4,12 +4,12 @@
 
 namespace Messages {
 
-    ClientDataMessage::ClientDataMessage(std::vector<GLfloat> position) {
+    ClientDataMessage::ClientDataMessage(std::vector<GLuint> position) {
         Type = Messages::MsgType::ClientData;
         ClientDataMessage::UpdatePosition(std::move(position));
     }
 
-    void ClientDataMessage::UpdatePosition(std::vector<GLfloat> position) {
+    void ClientDataMessage::UpdatePosition(std::vector<GLuint> position) {
         Position = std::move(position);
     }
 } // Messages
